tls fix - /ns and passwords flags

+++ Connection.js
@@ -1,44 +1,14 @@
 const net = require('net');
 const tls = require('tls');
-// const { ReStartTls } = require('./Helpers/ConnectionsPool');
-
 const Settings = require('./settings');
-
-const IrcParser = require("./Helpers/IrcParser");
-
-const {
-    WHOISUSER,
-    WHOISSERVER,
-    WHOISCHANNELS,
-    WHOISIDLE,
-    ENDOFWHOIS,
-    PING,
-    NICKNAMEINUSE,
-    CHANGEDNICK,
-    LEAVING,
-    JOIN,
-    GETNAMES,
-    HandleKill,
-    SHOWNAMES,
-    MOTD,
-    WELCOME,
-    SERVERINFO,
-    PRIVMSG,
-    SetTopic,
-    MODE,
-    ShowChanModes,
-    ShowChanDate,
-    ShowBanList,
-    ShowEndOfBanList,
-    AlreadyReg,
-    Notice,
-    Status
-} = require("./Controllers/index");
 const { ScreenUpdate } = require('./Helpers/ScreenUpdate');
+const { Listener } = require('./Listener');
 
-class Connecting {
-  constructor(server, port, user, realname, nickname, channels, tlsFlag) {
-    this.nickname = Settings[this.identity]?.nickname || nickname || Settings.nicknames[0];
+class Connection extends Listener {
+  constructor({server, port, user, realname, nickname, channels, password, serverpassword, tls, rejectUnauthorized}) {
+    super();
+    this.originalNick = nickname;
+    this.nickname = nickname || Settings[this.identity]?.nickname  || Settings.nicknames[0];
     this.server = server;
     this.port = port;
     this.channels = channels;
@@ -46,17 +16,44 @@ class Connecting {
     this.realname = realname || Settings.realname;
     this.connecting = false;
     this.preError = false;
-    this.tlsFlag = tlsFlag;
+    this.tlsFlag = tls;
+    this.password = password;
+    this.serverpassword = serverpassword;
     this.identity;
     this.options = {
-      rejectUnauthorized: false,
+      rejectUnauthorized: rejectUnauthorized,
       port: this.port,
       host: this.server
     };
   }
 
+  Start() {
+    this.MakeIdentity();
+    this.CreateSettings();
+    this.Connect();
+  }
+
   MakeIdentity() {
     this.identity = new Date().getTime() + Math.floor(Math.random() * 100);
+  }
+
+  CreateSettings() {
+    Settings[this.identity] = {
+      private: {},
+      identity: this.identity,
+      nickname: this.nickname,
+      server: this.server,
+      port: this.port,
+      channels: this.channels,
+      user: this.user,
+      realname: this.realname,
+      PassedMOTD: false,
+      count: 0,
+      joinedChans: [],
+      displayRaw : false,
+      raw : "",
+      status : ""
+    };
   }
 
   RebindClient(client){
@@ -64,43 +61,34 @@ class Connecting {
   }
 
   Connect() {
-    let client = new net.Socket();
-    const { Listen, identity , server } = this;
-    client.setTimeout(80000);
+    let client = null;
+    const { Listen, identity , server, tlsFlag, user, realname, serverpassword, password, originalNick } = this;
     const connect = () => {
-      if(this.tlsFlag && !this.preError){
+      if(this.tlsFlag){
         client = tls.connect(this.options);
+      } else {
+        client = net.connect(this.options); 
       }
-      if(this.tlsFlag && this.preError){
-        client = tls.connect(this.options);
-        BindEvents(client);
-        this.RebindClient(client);
-        ScreenUpdate({identity, client, server});
-      }
-      if(!this.tlsFlag){
-      client.connect(this.options);
-      }
-      
       client.setTimeout(80000);
       this.connecting = false;
-      Settings[this.identity].PassedMOTD = false;
-      clearTimeout(Settings[this.identity].timeout);
-      clearTimeout(Settings[this.identity].ping);
-      clearTimeout(Settings[this.identity].connect);
-      if(!this.preError){
-        BindEvents(client);
-        this.RebindClient(client);
-      }
+      Settings[identity].PassedMOTD = false;
+      clearTimeout(Settings[identity].ping);
+      clearTimeout(Settings[identity].connect);
       this.preError = false;
-
+      BindEvents(client);
+      this.RebindClient(client);
+      ScreenUpdate({identity, client, server});
+      this.UpdateScreen = true;
     };
 
     const connectHandler = () => {
+      let nickname = Settings[this.identity]?.nickname || this.nickname;
       client.write('CAP LS 302\r\n');
       client.write('CAP REQ multi-prefix\r\n');
       client.write('CAP END\r\n');
-      client.write(`NICK ${this.nickname}\r\n`);
-      client.write(`USER ${this.user} 0 * :${this.realname}\r\n`);
+      serverpassword ? client.write(`PASS ${serverpassword}\r\n`) : null;
+      client.write(`NICK ${nickname}\r\n`);
+      client.write(`USER ${user} 0 * :${realname}\r\n`);
       this.connecting = true;
       this.preError = false;
     };
@@ -113,7 +101,10 @@ class Connecting {
           Settings[this.identity][chan.toLowerCase()].chanNicks = [];
         });
       }
-      client.destroy();
+      if (client) {
+        client.destroy();
+        client = null;
+      }
       Settings[this.identity].connect = setTimeout(connect, 5000);
       Settings[this.identity].status += `^R${err} Reconnecting...^\r\n`;
     };
@@ -146,97 +137,19 @@ class Connecting {
     };
 
     function BindEvents(client){
-      console.logger('binding events!');
       client.on('error', errorHandler);
+      tlsFlag ? client.on('secureConnect', connectHandler) :
       client.on('connect', connectHandler);
       client.on('timeout', Timeout);
       client.on('close', Close);
-      client.on("data", (data) => Listen(data, identity, client, server));
+      client.on("data", (data) => Listen(data, identity, client, server, password, originalNick));
       
     }
 
     connect();
   }
 
-  CreateSettings() {
-    Settings[this.identity] = {
-      private: {},
-      identity: this.identity,
-      nickname: this.nickname,
-      server: this.server,
-      port: this.port,
-      channels: this.channels,
-      user: this.user,
-      realname: this.realname,
-      PassedMOTD: false,
-      count: 0,
-      joinedChans: [],
-      displayRaw : false,
-      raw : "",
-      status : ""
-    };
-  }
-
-
-  Listen(data, identity, client, server){
-    let message = [];
-    let temp = "";
-
-    data = data.toString();
-    let len = data.length - 1;
-    buffer(data);
-    function buffer(data) {
-        if (data[len] !== "\n") {
-            return temp += data;
-
-        }
-        if (data[len] === "\n") {
-            if (temp.length > 0) {
-                message.push(temp + data);
-                temp = "";
-                return;
-            }
-            if (data.length > 0) message.push(data);
-        }
-    }
-    const rnparse = message.flatMap((e) => e.split("\r\n"));
-    for (const raw of rnparse) {
-        if (raw.length > 0) {
-            const parsed = IrcParser(raw, identity);
-            const { command, params } = parsed;
-            Status(parsed);
-            if (command === "001") WELCOME(params, identity, client); // RPL_WELCOME (001)
-            if (command === "004") SERVERINFO(identity, server); // RPL_MYINFO (004) server info
-            if (command === "462") AlreadyReg(parsed, client);
-            if (command === "376" || command === "422") MOTD(client, parsed); // RPL_MOTD (372)
-            if (command === "433") NICKNAMEINUSE(client, parsed, identity); // ERR_NICKNAMEINUSE (433)
-            if (command === "ERROR") HandleKill(client, raw);
-            if (command === "NOTICE") Notice(parsed);
-            if (command === "PING" || command === "PONG") PING(parsed, params, client);
-            if (command === "353") GETNAMES(parsed); // RPL_NAMREPLY (353
-            if (command === "366") SHOWNAMES(parsed); // RPL_ENDOFNAMES (366)'
-            if (command === "NICK") CHANGEDNICK(parsed, identity);
-            if (command === "JOIN") JOIN(client, parsed);
-            if (command === "PART" || command === "QUIT" || command === "KICK") LEAVING(parsed, identity);
-            if (command === "311") WHOISUSER(params);
-            if (command === "312") WHOISSERVER(params, client);
-            if (command === "319") WHOISCHANNELS(params, client);
-            if (command === "317") WHOISIDLE(params, client);
-            if (command === "318") ENDOFWHOIS(parsed, client);
-            if (command === "332") SetTopic(parsed);
-            if (command === "PRIVMSG") PRIVMSG(parsed, client);
-            if (command === "MODE") MODE(parsed);
-            if(command === "324") ShowChanModes(parsed);
-            if(command === "329") ShowChanDate(parsed);
-            if(command === "367") ShowBanList(parsed);
-            if(command === "368") ShowEndOfBanList(parsed);
-            message.pop();
-        }
-    }
-}
-
-
 
 }
 
-module.exports = { Connecting };
+module.exports = { Connection };
